requirements.txt
anthropic>=0.40.0
click>=8.1.7
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
rich>=13.7.0
pandas>=2.2.0
numpy>=1.26.0
tweepy>=4.14.0
